Adds a bearer injecting auth guard for pdf file endpoint (#2325)
