+++ libs/judicial-system/auth/src/index.ts
@@ -2,6 +2,7 @@ export * from './lib/auth.module'
 export * from './lib/auth.service'
 export * from './lib/auth.types'
 export * from './lib/guards/auth.guard'
+export * from './lib/guards/auth.inject.bearer.guard'
 export * from './lib/guards/auth.graphql.guard'
 export * from './lib/guards/roles.guard'
 export * from './lib/decorators/graphql.user.decorator'
